Move AppStatesPreferenceCleaner to che-core-api-workspace.
