+++ assembly/assembly-wsmaster-war/src/main/java/org/eclipse/che/api/deploy/WsMasterModule.java
@@ -164,6 +164,6 @@ public class WsMasterModule extends AbstractModule {
 
     bind(ServerCheckerFactory.class).to(ServerCheckerFactoryImpl.class);
 
-    bind(AppStatesPreferenceCleaner.class);
+    bind(org.eclipse.che.api.workspace.server.AppStatesPreferenceCleaner.class);
   }
 }
